Multi-Class Deepfake/FSM-Net1/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64

# _libgcc_mutex=0.1=main  # This line can be removed as it's specific to Linux.
blas=1.0=mkl
bzip2=1.0.8=h8cc25b3_0
ca-certificates=2023.05.30=h5b45459_0
certifi=2021.5.30=py36haa95532_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cudatoolkit=11.1.74=h6bb024c_0
dataclasses=0.8=py_0
ffmpeg=4.3.1=h0cdd941_1
freetype=2.11.0=ha860e81_0
giflib=5.2.1=h8cc25b3_1
gmp=6.2.1=h74a9793_1
gnutls=3.6.15=he2a5e2d_0
intel-openmp=2022.0.1=ha5fd6cc_3634
jpeg=9e=h2bbff1b_0
lame=3.100=h1d878fa_0
lcms2=2.12=ha266a8f_0
# ld_impl_win-64=2.38=hea87114_0  # Windows-specific linker.
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.6.1.9=0
libcurand=10.3.2.106=0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libffi=3.3=he774522_2
# libgcc-ng=9.1.0=hdf63c60_0  # This line can be removed as it's specific to Linux.
libiconv=1.16=h1df5818_2
libidn2=2.3.2=hc5aeecc_0
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libpng=1.6.37=h2a8f88b_0
# libstdcxx-ng=9.1.0=hdf63c60_0  # This line can be removed as it's specific to Linux.
libtasn1=4.16.0=h3879d1d_0
libtiff=4.2.0=hd0e1b90_0
libunistring=0.9.10=h9b4dc7a_0
libuv=1.40.0=he774522_0
libwebp=1.2.2=hc3b3e54_0
libwebp-base=1.2.2=h2bbff1b_0
lz4-c=1.9.3=hf4a77e7_0
# ncurses=6.3=he774522_2  # Not required for Windows, can be removed.
nettle=3.7.3=hbbd107a_1
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
olefile=0.46=pyhd3eb1b0_0
openh264=2.1.1=h8ffe710_0
openssl=1.1.1t=h2bbff1b_0
pillow=8.3.1=py36hdc2b20a_0
pip=21.2.2=py36haa95532_0
python=3.6.13=h3758d61_1
pytorch=1.10.2=py3.6_cuda11.1_cudnn8_0
pytorch-cuda=11.7=h778d358_5
pytorch-mutex=1.0=cuda
# readline=8.1.2=h7f8727e_1  # Not required for Windows, can be removed.
setuptools=58.0.4=py36haa95532_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.5=h2bbff1b_0
tk=8.6.12=h2bbff1b_0
torchaudio=0.10.2=py36_cu111
torchvision=0.11.3=py36_cu111
typing_extensions=4.1.1=pyh06a4308_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h62dcd97_1
zlib=1.2.12=h62dcd97_2
zstd=1.5.2=h19a0ad4_0